The app is throwing an error when we start, please help us. Also, tell us what caused the issue.  --Done
The rental class has a method to save, but it is not async, can you make it async and explain to us what is the difference?

Please finish the method to filter rentals by customer name, and add the new endpoint.

We noticed we do not have a table for customers, it is not good to have just the customer name in the rental. Can you help us add a new entity for this? Don't forget to change the customer name field to a foreign key, and fix your previous method!

In the MovieFeatures class, there is a method to list all movies, tell us your opinion about it.

No exceptions are being caught in this api, how would you deal with these exceptions?

Challenge (Nice to have)
--TODO